use vitest 3.2.6 for blackbox for consistency
