+++ tests/blackbox/tests/db/routes/items/m2o.test.ts
@@ -7,7 +7,7 @@ import type { PrimaryKeyType } from '@common/types';
 import { PRIMARY_KEY_TYPES, USER } from '@common/variables';
 import { without } from 'lodash-es';
 import request from 'supertest';
-import { beforeAll, describe, expect, it } from 'vitest';
+import { beforeAll, describe, expect, it, test } from 'vitest';
 import { type CachedTestsSchema, CheckQueryFilters, type TestsSchemaVendorValues } from '../../query/filter';
 import {
 	type City,
@@ -2032,82 +2032,81 @@ describe.each(PRIMARY_KEY_TYPES)('/items', (pkType) => {
 			});
 		});
 
-		describe.runIf(pkType === 'integer' && without(vendors, 'cockroachdb', 'mssql', 'oracle').length > 0)(
-			'Auto Increment Tests',
-			() => {
-				describe('updates the auto increment value correctly', () => {
-					it.each(without(vendors, 'cockroachdb', 'mssql', 'oracle'))('%s', async (vendor) => {
-						// Setup
-						const name = 'test-auto-increment-m2o';
-						const largeIdCity = 102222;
-						const largeIdState = 103333;
-						const largeIdCountry = 104444;
-						const city = createCity(pkType);
-						const city2 = createCity(pkType);
-						const state = createState(pkType);
-						const state2 = createState(pkType);
-						const country = createCountry(pkType);
-						const country2 = createCountry(pkType);
+		test('Auto Increment Tests', (ctx) => {
+			if (pkType !== 'integer') ctx.skip();
 
-						city.id = largeIdCity;
-						state.id = largeIdState;
-						country.id = largeIdCountry;
-						city.name = name;
-						city2.name = name;
-						state.name = name;
-						state2.name = name;
-						country.name = name;
-						country2.name = name;
+			describe('updates the auto increment value correctly', () => {
+				it.each(without(vendors, 'cockroachdb', 'mssql', 'oracle'))('%s', async (vendor) => {
+					// Setup
+					const name = 'test-auto-increment-m2o';
+					const largeIdCity = 102222;
+					const largeIdState = 103333;
+					const largeIdCountry = 104444;
+					const city = createCity(pkType);
+					const city2 = createCity(pkType);
+					const state = createState(pkType);
+					const state2 = createState(pkType);
+					const country = createCountry(pkType);
+					const country2 = createCountry(pkType);
 
-						await CreateItem(vendor, {
-							collection: localCollectionCities,
-							item: [
-								{
-									...city,
-									state_id: {
-										...state,
-										country_id: country,
-									},
+					city.id = largeIdCity;
+					state.id = largeIdState;
+					country.id = largeIdCountry;
+					city.name = name;
+					city2.name = name;
+					state.name = name;
+					state2.name = name;
+					country.name = name;
+					country2.name = name;
+
+					await CreateItem(vendor, {
+						collection: localCollectionCities,
+						item: [
+							{
+								...city,
+								state_id: {
+									...state,
+									country_id: country,
 								},
-								{
-									...city2,
-									state_id: {
-										...state2,
-										country_id: country2,
-									},
+							},
+							{
+								...city2,
+								state_id: {
+									...state2,
+									country_id: country2,
 								},
-							],
-						});
-
-						// Action
-						const response = await request(getUrl(vendor))
-							.get(`/items/${localCollectionCities}`)
-							.query({
-								filter: JSON.stringify({
-									name: { _eq: name },
-								}),
-								fields: 'id,state_id.id,state_id.country_id.id',
-							})
-							.set('Authorization', `Bearer ${USER.ADMIN.TOKEN}`);
-
-						// Assert
-						expect(response.statusCode).toBe(200);
-						expect(response.body.data.length).toBe(2);
-
-						expect(response.body.data.map((city: any) => city.id)).toEqual(
-							Array.from({ length: 2 }, (_, index) => largeIdCity + index),
-						);
-
-						expect(response.body.data.map((city: any) => city.state_id.id)).toEqual(
-							Array.from({ length: 2 }, (_, index) => largeIdState + index),
-						);
-
-						expect(response.body.data.map((city: any) => city.state_id.country_id.id)).toEqual(
-							Array.from({ length: 2 }, (_, index) => largeIdCountry + index),
-						);
+							},
+						],
 					});
+
+					// Action
+					const response = await request(getUrl(vendor))
+						.get(`/items/${localCollectionCities}`)
+						.query({
+							filter: JSON.stringify({
+								name: { _eq: name },
+							}),
+							fields: 'id,state_id.id,state_id.country_id.id',
+						})
+						.set('Authorization', `Bearer ${USER.ADMIN.TOKEN}`);
+
+					// Assert
+					expect(response.statusCode).toBe(200);
+					expect(response.body.data.length).toBe(2);
+
+					expect(response.body.data.map((city: any) => city.id)).toEqual(
+						Array.from({ length: 2 }, (_, index) => largeIdCity + index),
+					);
+
+					expect(response.body.data.map((city: any) => city.state_id.id)).toEqual(
+						Array.from({ length: 2 }, (_, index) => largeIdState + index),
+					);
+
+					expect(response.body.data.map((city: any) => city.state_id.country_id.id)).toEqual(
+						Array.from({ length: 2 }, (_, index) => largeIdCountry + index),
+					);
 				});
-			},
-		);
+			});
+		});
 	});
 });
