+++ tests/blackbox/vitest.config.ts
@@ -5,37 +5,16 @@ import Sequencer from './setup/sequencer';
 export default defineConfig({
 	plugins: [tsconfigPaths()],
 	test: {
-		pool: 'forks',
-		maxWorkers: 6,
+		poolOptions: {
+			forks: {
+				minForks: 1,
+				maxForks: 6,
+			},
+		},
+		environment: './setup/environment.ts',
 		sequence: {
 			sequencer: Sequencer,
 		},
 		testTimeout: 15_000,
-		projects: [
-			{
-				extends: true,
-				test: {
-					name: 'common',
-					include: ['tests/common/**/*.test.ts', 'common/common.test.ts'],
-					globalSetup: './setup/setup.ts',
-					setupFiles: ['./setup/sequential-gate.ts'],
-					env: {
-						TEST_PROJECT: 'common',
-					},
-				},
-			},
-			{
-				extends: true,
-				test: {
-					name: 'db',
-					include: ['tests/db/**/*.test.ts', 'common/common.test.ts'],
-					globalSetup: './setup/setup.ts',
-					setupFiles: ['./setup/sequential-gate.ts'],
-					env: {
-						TEST_PROJECT: 'db',
-					},
-				},
-			},
-		],
 	},
 });
